bunch of micro optimizations

+++ Interactive-Evolution-SFX/src/SoundData.cpp
@@ -65,5 +65,5 @@ void IESFX::SoundData::poke(RESID::reg8 offset, RESID::reg8 value)
 void SoundData::sample(size_t size)
 {
 	RESID::cycle_count delta_t = util::get_cycles(util::get_size(size));
-	_index += static_cast<size_t>(_sid->clock(delta_t, _samples.data() + _index, util::get_size(size)));
+	_index += _sid->clock(delta_t, _samples.data() + _index, util::get_size(size));
 }
